Extract DtaOrchestrator and wire plugin through it (#23)
DtaOrchestrator in dta-daemon is the shared brain that manages
connections, CDP watchers, SSE events, layout tree enrichment,
and data fetching.

- DtaOrchestrator: singleton with all orchestration logic extracted
  from InspectorController
- InspectorController: thinned to pure REST adapter
- DtaService (plugin): uses DtaOrchestrator for full capabilities
  including CDP capture and layout enrichment
